I tried, and ultimately failed, to recreate IRAF's geotran using numpy.
As it turns out, the mathematics behind geotran have been lost to the sands of
time. According to iraf.net, the first line for any surface in an IRAF
Fitcoords .dbs file is the function-type: 1 for Chebyshev, or 2 for Legrende.
The .dbs files generated by geotran have a 3.
*clap* *clap* *clap*
Whoever was working on geotran last century, when IRAF was still used, added
another function code, and didn't have the common sense to update the
documentation page. Which means we are now incapable of performing any
transformations with orders of more than 2. I tried. I was off by an order
of magnitude. Oh, well. I guess I'll just revert this, since it doesn't really
work. I tried using a polynomial, but apparently that made too much goddarned
sense.